[Incomplete] Added highlight.js to parse code blocks.
I need help with how to add the styles highlight.js, at the moment I just added it to react.tsx but this is not the right place. If you show me how to add this properly I will do it! Also I need to know if this approach is going to work on the React Native app? (since it uses divs, will this work?) comments would be very much appreciated!